6th commit

+++ cinema/src/main/java/com/mytix/cinema/repository/GenreRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.mytix.cinema.model.Genre;
 
 public interface GenreRepository 
-extends JpaRepository<Genre, Long>{
+extends JpaRepository<Genre, Integer>{
     
     @Query("select g from Genre g where g.genreName=:genreName")
     Optional<Genre> findByGenreName(String genreName);
